// iam1.client/src/models/UserProfile.ts
export interface UserProfile {
    fullName: string;
    employeeId: string;
    gender: string;
    dateOfBirth: string;
    email: string;
    phone: string;
    officeLocation: string;
    department: string;
    position: string;
    managerEmail: string;
    joinedDate: string;
    employeeStatus: string;
    employeeType: string;
}